nmrscripts/predistancefinder/requirements.txt
pandas
scipy
matplotlib
sympy
biopython
